# docs/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: 筑享开发团队
  description: 筑享移动端和管理端后端接口。除明确标注为公开接口外，受保护接口需在 Authorization 请求头中携带 Bearer 访问令牌。
  license:
    name: 内部使用
  title: 筑享租房平台接口文档
  version: v1
servers:
- description: Generated server url
  url: http://localhost:18080/api
...
paths:
...
  /houses:
    get:
      description: 按关键字、分类、区域、价格、户型、面积和设施组合筛选房源，并支持排序和分页。
      operationId: search
      parameters:
      - description: 标题、位置或小区关键字
        in: query
        name: keyword
        required: false
        schema:
          type: string
      - description: 房源分类键
        in: query
        name: category
        required: false
        schema:
          type: string
      - description: 区域名称
        in: query
        name: region
        required: false
        schema:
          type: string
      - description: 最低月租金，单位元
        example: 1000
        in: query
        name: minPrice
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - description: 最高月租金，单位元
        example: 5000
        in: query
        name: maxPrice
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - description: 户型
        in: query
        name: roomType
        required: false
        schema:
          type: string
      - description: 最小面积，单位平方米
        example: 20
        in: query
        name: minArea
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - description: 最大面积，单位平方米
        example: 120
        in: query
        name: maxArea
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - description: 设施值，多个值使用英文逗号分隔
        example: "wifi,air_conditioner"
        in: query
        name: facilities
        required: false
        schema:
          type: string
      - description: 排序值，以筛选选项接口返回值为准
        example: default
        in: query
        name: sort
        required: false
        schema:
          type: string
          default: default
      - description: 页码，从 1 开始
        example: 1
        in: query
        name: page
        required: false
        schema:
          type: integer
          format: int64
          default: 1
          minimum: 1
      - description: 每页条数，范围 1-100
        example: 20
        in: query
        name: pageSize
        required: false
        schema:
          type: integer
          format: int64
          default: 20
          maximum: 100
          minimum: 1
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePageDataHouseView"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      summary: 搜索房源
      tags:
      - 房源
  /houses/feed:
    get:
      description: 按首页分类分页获取房源和插入广告；携带令牌时返回收藏状态。
      operationId: feed
      parameters:
      - description: 首页分类键，以首页 tabs 返回值为准
        example: recommend
        in: query
        name: category
        required: true
        schema:
          type: string
      - description: 页码，从 1 开始
        example: 1
        in: query
        name: page
        required: false
        schema:
          type: integer
          format: int64
          default: 1
          minimum: 1
      - description: 每页条数，范围 1-100
        example: 20
        in: query
        name: pageSize
        required: false
        schema:
          type: integer
          format: int64
          default: 20
          maximum: 100
          minimum: 1
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseFeedData"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      summary: 获取房源流
      tags:
      - 房源
  /houses/filter-options:
    get:
      description: 返回区域、价格区间、户型、设施和排序选项。
      operationId: filterOptions
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseFilterOptions"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      summary: 获取筛选选项
      tags:
      - 房源
  /houses/{houseId}:
    get:
      description: 返回房源图片、价格、地址、配套、房东资料和出租状态；携带令牌时返回收藏状态。
      operationId: detail_1
      parameters:
      - description: 房源 ID
        example: house_001
        in: path
        name: houseId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseHouseDetail"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      summary: 获取房源详情
      tags:
      - 房源
  /houses/{houseId}/favorite:
    delete:
      description: 从当前用户收藏中移除指定房源。
      operationId: unfavorite
      parameters:
      - description: 房源 ID
        example: house_001
        in: path
        name: houseId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseFavoriteResult"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 取消收藏房源
      tags:
      - 房源
    post:
      description: 将指定房源加入当前用户收藏。
      operationId: favorite
      parameters:
      - description: 房源 ID
        example: house_001
        in: path
        name: houseId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseFavoriteResult"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 收藏房源
      tags:
      - 房源
...
components:
  schemas:
...
    AdvertisementView:
      type: object
      description: 房源流广告
      properties:
        description:
          type: string
          description: 详细说明
        id:
          type: string
          description: 记录 ID
        imageUrl:
          type: string
          description: 图片 URL
        targetType:
          type: string
          description: 跳转目标类型
        targetValue:
          type: string
          description: 跳转目标值
        title:
          type: string
          description: 标题
...
    ApiResponseFavoriteResult:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          $ref: "#/components/schemas/FavoriteResult"
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
    ApiResponseFeedData:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          $ref: "#/components/schemas/FeedData"
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
...
    ApiResponseFilterOptions:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          $ref: "#/components/schemas/FilterOptions"
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
...
    ApiResponseHouseDetail:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          $ref: "#/components/schemas/HouseDetail"
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
...
    ApiResponsePageDataHouseView:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          $ref: "#/components/schemas/PageDataHouseView"
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
...
    ApiResponseVoid:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
...
    FavoriteResult:
      type: object
      description: 房源收藏操作结果
      properties:
        houseId:
          type: string
          description: 房源 ID
        isFavorite:
          type: boolean
          description: 当前用户是否已收藏
    FeedData:
      type: object
      description: 房源流分页数据
      properties:
        hasMore:
          type: boolean
          description: 是否还有下一页
        items:
          type: array
          description: 当前页条目
          items:
            $ref: "#/components/schemas/FeedItem"
        page:
          type: integer
          format: int64
          description: 当前页码，从 1 开始
        pageSize:
          type: integer
          format: int64
          description: 每页条数
    FeedItem:
      type: object
      description: 房源流条目，可为房源或广告
      properties:
        advertisement:
          $ref: "#/components/schemas/AdvertisementView"
          description: 广告数据；条目为房源时为空
        house:
          $ref: "#/components/schemas/HouseView"
          description: 房源数据；条目为广告时为空
        type:
          type: string
          description: 条目类型
...
    FilterOptions:
      type: object
      description: 房源搜索筛选选项
      properties:
        facilities:
          type: array
          description: 配套设施
          items:
            $ref: "#/components/schemas/Option"
        priceRanges:
          type: array
          description: 租金区间选项
          items:
            $ref: "#/components/schemas/PriceRange"
        regions:
          type: array
          description: 区域选项
          items:
            $ref: "#/components/schemas/Option"
        roomTypes:
          type: array
          description: 户型选项
          items:
            $ref: "#/components/schemas/Option"
        sortOptions:
          type: array
          description: 排序选项
          items:
            $ref: "#/components/schemas/Option"
...
    HouseDetail:
      type: object
      description: 房源完整详情
      properties:
        address:
          type: string
          description: 详细地址
        area:
          type: integer
          format: int32
          description: 建筑面积，单位平方米
        availableDate:
          type: string
          format: date
          description: 最早可入住日期
        avatarUrl:
          type: string
          description: 头像图片 URL
        community:
          type: string
          description: 小区名称
        coverImage:
          type: string
          description: 封面图片 URL
        decoration:
          type: string
          description: 装修情况
        deposit:
          type: integer
          format: int32
          description: 押金，单位元
        description:
          type: string
          description: 详细说明
        facilities:
          type: array
          description: 配套设施
          items:
            type: string
        floor:
          type: string
          description: 楼层描述
        id:
          type: string
          description: 记录 ID
        images:
          type: array
          description: 房源图片列表
          items:
            type: string
        isFavorite:
          type: boolean
          description: 当前用户是否已收藏
        isSmartLockSupported:
          type: boolean
          description: 是否支持智能门锁
        isVerified:
          type: boolean
          description: 是否已认证
        landlordId:
          type: string
          description: 房东用户 ID
        landlordName:
          type: string
          description: 房东姓名
        location:
          type: string
          description: 区域或商圈位置
        metro:
          type: string
          description: 附近地铁信息
        orientation:
          type: string
          description: 房屋朝向
        paymentMethod:
          type: string
          description: 租金付款方式
        price:
          type: integer
          format: int32
          description: 月租金，单位元
        rating:
          type: number
          description: 综合评分
        rentedCount:
          type: integer
          format: int32
          description: 累计出租次数
        responseDescription:
          type: string
          description: 响应速度说明
        roomType:
          type: string
          description: 户型
        status:
          type: string
          description: 业务状态
        tags:
          type: array
          description: 房源标签
          items:
            type: string
        title:
          type: string
          description: 标题
    HouseView:
      type: object
      description: 房源列表摘要
      properties:
        area:
          type: integer
          format: int32
          description: 建筑面积，单位平方米
        availableDate:
          type: string
          format: date
          description: 最早可入住日期
        community:
          type: string
          description: 小区名称
        coverImage:
          type: string
          description: 封面图片 URL
        decoration:
          type: string
          description: 装修情况
        description:
          type: string
          description: 详细说明
        facilities:
          type: array
          description: 配套设施
          items:
            type: string
        floor:
          type: string
          description: 楼层描述
        id:
          type: string
          description: 记录 ID
        isFavorite:
          type: boolean
          description: 当前用户是否已收藏
        isSmartLockSupported:
          type: boolean
          description: 是否支持智能门锁
        location:
          type: string
          description: 区域或商圈位置
        metro:
          type: string
          description: 附近地铁信息
        orientation:
          type: string
          description: 房屋朝向
        price:
          type: integer
          format: int32
          description: 月租金，单位元
        roomType:
          type: string
          description: 户型
        status:
          type: string
          description: 业务状态
        tags:
          type: array
          description: 房源标签
          items:
            type: string
        title:
          type: string
          description: 标题
...
    Option:
      type: object
      description: 通用筛选选项
      properties:
        label:
          type: string
          description: 选项显示名称
        value:
          type: string
          description: 选项提交值
    PageDataHouseView:
      type: object
      description: 分页数据
      properties:
        hasMore:
          type: boolean
          description: 是否还有下一页
          example: true
        items:
          type: array
          description: 当前页数据
          items:
            $ref: "#/components/schemas/HouseView"
        page:
          type: integer
          format: int64
          description: 当前页码，从 1 开始
          example: 1
        pageSize:
          type: integer
          format: int64
          description: 每页条数
          example: 20
        total:
          type: integer
          format: int64
          description: 总记录数
          example: 42
...
    PriceRange:
      type: object
      description: 租金筛选区间
      properties:
        label:
          type: string
          description: 选项显示名称
        maxPrice:
          type: integer
          format: int32
          description: 最高月租金，单位元
        minPrice:
          type: integer
          format: int32
          description: 最低月租金，单位元
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: "访问令牌，格式：Bearer {accessToken}"
      name: Authorization
      scheme: bearer
      type: http